refactor(arika): move kepler module from orts to arika
KeplerianElements and the Kepler-equation / anomaly conversion helpers are
fundamental astrodynamics primitives that belong alongside coordinate frames
and epochs in arika, not in the orts platform crate. This is the enabling
step for placing element-set parsers (TLE / OMM) in arika so they can produce
frame-typed coordinates directly.

- New arika::kepler module, adapted to no_std (core::f64::consts::PI +
  crate::math::F64Ext so f64 transcendentals resolve via libm).
- orts::orbital::kepler re-exports arika::kepler::* to keep the historical
  path working; no call sites change.
